Implemented different file structure
