+++ app.js
@@ -1,10 +1,10 @@
-const server = require('./server');
-const mongoose = require('mongoose');
-const port = process.env.PORT || 8008;
+// const server = require('./server');
+// const mongoose = require('mongoose');
+// const port = process.env.PORT || 8008;
 
-mongoose
-    .connect('mongodb://admin:admin@ds141320.mlab.com:41320/lambda-notes')
-    .then(() => console.log('mLab Connected to MongoDB'))
-    .catch(() => console.log('Error Connecting to mLab DB'))
+// mongoose
+//     .connect('mongodb://admin:admin@ds141320.mlab.com:41320/lambda-notes')
+//     .then(() => console.log('mLab Connected to MongoDB'))
+//     .catch(() => console.log('Error Connecting to mLab DB'))
 
-server.listen(port, () => console.log(`Server Connected to ${port}`));
+// server.listen(port, () => console.log(`Server Connected to ${port}`));
